Rename initial data generation module for clarity
- Rename `controller_operation.py` to `initial_data_generation.py`
  to better reflect its purpose.
- Update import statements in example files.
