+++ examples/lti_control/robust_lti_dd_mpc_reproduction.py
@@ -42,7 +42,7 @@ import os
 
 from utilities.controller.controller_params import (
     get_lti_data_driven_mpc_controller_params)
-from utilities.controller.controller_operation import (
+from utilities.controller.initial_data_generation import (
     randomize_initial_system_state, simulate_n_input_output_measurements,
     generate_initial_input_output_data)
 from utilities.reproduction.paper_reproduction import (
